move text formatter to a separate package
